+++ aihub_api/aihub_api/routes/evaluation/EvaluationService.py
@@ -7,7 +7,7 @@ import httpx
 import pandas as pd
 import phoenix as px
 from aihub_lib.auth.AuthenticatedUser import AuthenticatedUser
-from aihub_lib.generative_ai.evaluation.Evaluator import PhoenixExperimentEvaluator
+from aihub_lib.generative_ai.evaluation.PhoenixExperimentEvaluator import PhoenixExperimentEvaluator
 from aihub_lib.generative_ai.resources.models.llm.chat.ChatLLMConfig import ChatLLMConfig
 from aihub_lib.i18n.LocaleHandler import LocaleHandler
 from aihub_lib.infrastructure.phoenix.PhoenixConfig import PhoenixConfig
