+++ aihub_api/aihub_api/routes/evaluation/dto/experiment/Experiment.py
@@ -53,4 +53,4 @@ class Experiment(MinimalExperiment):
         Field(
             description="Detailed records of each run within the experiment, including inputs, outputs, and evaluations."
         ),
-    ] = None
+    ]
