+++ aihub_pipeline/pyproject.toml
@@ -20,9 +20,8 @@ matplotlib = "^3.9.2"
 adlfs = "^2024.12.0"
 
 [tool.poetry.dependencies.aihub_lib]
-git = "https://github.com/bbvch-ai/aihub-core.git"
-tag = "v0.168.0"
-subdirectory = "aihub_lib"
+path = "../aihub_lib"
+develop = true
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.2"
