+++ aihub_web/aihub_web/components/Evaluation/Dataset/Create.vue
@@ -74,7 +74,3 @@ const save = () => {
   emit('close')
 }
 </script>
-
-<style scoped>
-
-</style>
