+++ aihub_web/aihub_web/components/Knowledge/Node/Content.vue
@@ -69,7 +69,3 @@ const scoreRank = computed<string>(() => {
   return 'A++'
 })
 </script>
-
-<style scoped>
-
-</style>
