+++ aihub_web/aihub_web/pages/admin/evaluation.vue
@@ -56,7 +56,3 @@ const activeNavItem = computed<NavItem | undefined>(() => {
   return navItems.value?.filter(navItem => navItem.isActive())[0]
 })
 </script>
-
-<style scoped>
-
-</style>
