+++ aihub_web/aihub_web/pages/admin/evaluation/dataset/[dataset_id].vue
@@ -76,7 +76,3 @@ const safeDataset = (event) => {
   })
 }
 </script>
-
-<style scoped>
-
-</style>
