+++ aihub_web/aihub_web/pages/admin/evaluation/experiment/[experiment_id].vue
@@ -12,7 +12,3 @@
 <script setup lang="ts">
 const { experiment, experimentIsLoading } = useExperiment()
 </script>
-
-<style scoped>
-
-</style>
