+++ aihub_web/aihub_web/pages/admin/knowledge/[db]/[namespace]/[document_id]/overview.vue
@@ -18,7 +18,3 @@
 const { document, documentIsLoading } = useDocument()
 const route = useRoute()
 </script>
-
-<style scoped>
-
-</style>
