+++ aihub_web/aihub_web/pages/service/openai/[thread_id]/[display_id]/sources.vue
@@ -93,7 +93,3 @@ const documentMap = computed<Record<string, DocumentInfo>>(() => {
   return docs
 })
 </script>
-
-<style scoped>
-
-</style>
